Khang29/12
Khang:
*Bổ sung:
     -Opening scene
     -Ending scene
     -RollOut scene
*Cập nhật:
     -Thêm tính năng chọn đạn chưa hoàn thiện.
     -FixBug một số chức năng.

+++ BlasterMasterNES/GameObject.h
@@ -52,7 +52,7 @@ public:
 	float vx;
 	float vy;
 
-	int nx, ny;
+	float nx, ny;
 	
 	int state;
